+++ tavla/app/(admin)/hooks/useOrganizations.ts
@@ -2,7 +2,7 @@ import { NormalizedDropdownItemType } from '@entur/dropdown'
 import { useCallback, useEffect, useState } from 'react'
 import { getOrganizationsForUser } from '../actions'
 import { TOrganization } from 'types/settings'
-import { organizationToDropdownItem } from '../tavler/[id]/rediger/utils'
+import { organizationToDropdownItem } from '../tavler/[id]/utils'
 
 function useOrganizations(organization?: TOrganization) {
     const [organizationList, setOrganizationList] = useState<
