+++ tavla/app/(admin)/tavler/[id]/utils.ts
